requirements.txt
# Core
python-dotenv>=0.19.0
watchdog>=2.1.6
python-magic>=0.4.24

# Document Processing
pypdf2>=2.10.0
python-docx>=0.8.11
python-pptx>=0.6.21
pdf2image>=1.16.0
pytesseract>=0.3.8
pillow>=9.0.0
extract-msg>=0.28.7
email-validator>=1.1.3
docx2txt>=0.8
unstructured>=0.10.0

# Database
sqlalchemy>=1.4.36
alembic>=1.7.5

# AI/ML
spacy>=3.2.0
langchain>=0.0.267
openai>=0.27.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4  # For vector similarity search

# Web Interface
fastapi>=0.75.0
uvicorn>=0.17.6
python-multipart>=0.0.5

# Utilities
python-dateutil>=2.8.2
pytz>=2021.3
tqdm>=4.62.3

# Text Processing
nltk>=3.8.1

# For better PDF processing
pypdf>=3.0.0
pdfminer.six>=20221105

# For document embeddings
sentence-transformers>=2.2.2
